requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
SQLAlchemy==1.4.34
PyYAML==6.0
pymysql==1.0.2

pandas==1.4.2
botocore==1.15.32
boto3==1.12.32
s3fs==0.5.1
fsspec==0.8.4

joblib==1.1.0
matplotlib==3.5.1
numpy==1.22.3
scikit-learn==1.1.1